usm: Move HTTP replace rules to USM config
Move ReplaceRule struct and parseReplaceRules function from
replace_rules.go into usm_config.go for better organization.
